refactor: Modularize navigation logic. Split parts into components.
