+++ app/src/main/java/no/usn/mob3000/ui/screens/auth/CreateUserScreen.kt
@@ -25,7 +25,7 @@ import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.flow.Flow
 import no.usn.mob3000.R
-import no.usn.mob3000.Viewport
+import no.usn.mob3000.ui.components.base.Viewport
 import no.usn.mob3000.domain.model.auth.error.RegistrationError
 import no.usn.mob3000.domain.model.auth.state.RegistrationState
 import no.usn.mob3000.ui.components.Loading
