+++ app/src/main/java/no/usn/mob3000/ui/screens/auth/ForgotPasswordScreen.kt
@@ -11,7 +11,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import no.usn.mob3000.Viewport
+import no.usn.mob3000.ui.components.base.Viewport
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.height
