+++ app/src/main/java/no/usn/mob3000/ui/screens/auth/LoginScreen.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.unit.dp
-import no.usn.mob3000.Viewport
+import no.usn.mob3000.ui.components.base.Viewport
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.*
@@ -28,7 +28,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.font.FontWeight
 import kotlinx.coroutines.flow.Flow
-import no.usn.mob3000.Destination
+import no.usn.mob3000.domain.enumerate.Destination
 import no.usn.mob3000.R
 import no.usn.mob3000.domain.model.auth.state.LoginState
 import no.usn.mob3000.ui.components.Loading
