+++ app/src/main/java/no/usn/mob3000/ui/screens/info/AboutUsScreen.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import no.usn.mob3000.R
-import no.usn.mob3000.Viewport
+import no.usn.mob3000.ui.components.base.Viewport
 
 /**
  * Screen that shows details about the company.
