+++ app/src/main/java/no/usn/mob3000/ui/screens/info/docs/DocumentationDetailsScreen.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import no.usn.mob3000.R
-import no.usn.mob3000.Viewport
+import no.usn.mob3000.ui.components.base.Viewport
 import java.text.SimpleDateFormat
 import java.util.*
